refactor: `src/tui/event_handler.py`
