+++ src/tui/event_processing/__init__.py
@@ -1,0 +1,7 @@
+"""Internal event processing package for the TUI."""
+
+from .processor import EventProcessor
+from .stream import AsyncEventStream
+from .file_watcher import EventFileWatcher
+
+__all__ = ["EventProcessor", "AsyncEventStream", "EventFileWatcher"]
