+++ src/tui/event_processing/logging_config.py
@@ -1,0 +1,9 @@
+"""Logging utilities shared across event processing modules."""
+
+import logging
+
+# Preserve the original logger name for consistent output
+LOGGER_NAME = "src.tui.event_handler"
+logger = logging.getLogger(LOGGER_NAME)
+
+__all__ = ["logger", "LOGGER_NAME"]
